Read all the available data if any from the command line argument into array datastructure.
Added check for handling empty command line argument.

If data is not available empty string is printed on console.

else declare and initialize hashmap that stores the photogenic information.

Iterate over input array from index location 2 because data is available
from that location onwards.

Then checked if integer in string format is not a single digit number run another loop
over that integer and, perform mapping for each digit. Once complete
append it to the string output.

If single digit integer appears in the input array, output returned by hashmap
lookup is appended to the string output.

Added logic for comma insertion till the second last element

Finally the output that was stored in the string is printed on the
console.